requirements.txt
aiohttp
python-dotenv
mistralai
